[fix] validation function esm 방식으로 변경
